requirements.txt
fastapi
uvicorn 
sqlmodel
mysqlclient
python-jose[cryptography]
passlib[bcrypt]